requirements.txt
streamlit
fpdf
supabase
langchain-groq
langchain-community
langchain-core
langgraph
tavily-python
pandas
openpyxl
matplotlib
seaborn
tabulate
langchain_experimental
statsmodels
scikit-learn
streamlit-mic-recorder
fpdf
audio-recorder-streamlit